backend/requirements.txt
Django>=4.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
mysqlclient>=2.2
PyMuPDF>=1.23
python-dotenv>=1.0
requests>=2.31
Pillow>=10.0
